// WADNR.Web/src/app/shared/generated/model/location-simple-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class LocationSimpleStepRequest { 
    Latitude?: number;
    Longitude?: number;
    ProjectLocationSimpleTypeID?: number;
    ProjectLocationNotes?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface LocationSimpleStepRequestForm { 
    Latitude?: FormControl<number>;
    Longitude?: FormControl<number>;
    ProjectLocationSimpleTypeID?: FormControl<number>;
    ProjectLocationNotes?: FormControl<string>;
}

export class LocationSimpleStepRequestFormControls { 
    public static Latitude = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Longitude = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectLocationSimpleTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectLocationNotes = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}